services/_templates/fast_api/src/pyproject.toml
[tool.poetry]
name = "your-project-name"
version = "0.1.0"
description = "YOUR SITE API"
authors = ["You <you.com>"]
license = "All Rights Reserved"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
uvicorn = "^0.20.0"
fastapi = "^0.88.0"
sqlmodel = "^0.0.8"
redis = "^4.4.0"
oauthlib = "^3.2.2"
requests-oauthlib = "^1.3.1"
google-cloud-secret-manager = "^2.12.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.10.1"
black = "^22.10.0"
psycopg2-binary = "^2.9.5"
alembic = "^1.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
per-file-ignores = {"__init__.py" = ["F401"]}